Change number of generations without losing your place
When the number of generations is changed, focus the current field so it
is brought into view and the user doesn't have to go looking for it. If
that field is gone, the individual's nearest visible descendant will be
focused. If a "More" indicator is clicked, the individual next to it
will be focused. As a last resort, focus the offspring.