Improve breakdown of result
Previously, each row in the breakdown specified an ancestor, an
inbreeding coefficient, and the number of paths through that ancestor
with that coefficient. Thus there could be multiple rows per ancestor.

Now, there is one row per common ancestor. It includes the total
inbreeding coefficient of all paths through the ancestor, and the list
of individual paths with their coefficients.