Include the most recent change when showing offspring's data
If the user changes a field and then clicks "Show offspring's data"
while the field is still focused, the data is shown before the change is
incorporated. Fixed by triggering the change event on the field before
showing the data. This was already in place for the "Show selected
individual's data" button.