Added licensing
The Inbreeding calculator is now licensed under the GPLv3. The About
page is licensed under CC-BY-NC-ND. All files contain licensing
information. The new file instructions.html contains the instructions
portion of about.html and is licensed under the GPL. The new file
LICENSE is the GPL. The new file README explains this stuff. Deleted a
file which was committed by accident.